introduce .easy_command()
the command/push/execute dance got old fairly quickly, so I thought that the
API could benefit from a simple way to execute a command with a single method.

The new test plagiarizes the timeout vtc, feel free to call me out on
that one if you think it needs more coverage.

I wanted to keep the initial commit dead simple, so I haven't tried
supporting the `timeout` or `master` arguments, which should be easy to
do if we accept putting them as first optional arguments.